tests folder moved to core, all imports changed to relative imports
