+++ src/app/api/v1/login.py
@@ -1,16 +1,16 @@
 from typing import Annotated, Dict
-from datetime import timedelta, datetime, timezone
+from datetime import timedelta
 
 from fastapi import Response, Request, Depends
 from fastapi.security import OAuth2PasswordRequestForm
 from sqlalchemy.ext.asyncio import AsyncSession
 import fastapi
 
-from app.core.config import settings
-from app.core.db.database import async_get_db
-from app.core.exceptions.http_exceptions import UnauthorizedException
-from app.core.schemas import Token
-from app.core.security import (
+from ...core.config import settings
+from ...core.db.database import async_get_db
+from ...core.exceptions.http_exceptions import UnauthorizedException
+from ...core.schemas import Token
+from ...core.security import (
     ACCESS_TOKEN_EXPIRE_MINUTES, 
     create_access_token, 
     authenticate_user,
