+++ src/app/api/v1/logout.py
@@ -4,9 +4,9 @@ from fastapi import APIRouter, Response, Depends
 from sqlalchemy.ext.asyncio import AsyncSession
 from jose import JWTError
 
-from app.core.security import oauth2_scheme, blacklist_token
-from app.core.db.database import async_get_db
-from app.core.exceptions.http_exceptions import UnauthorizedException
+from ...core.security import oauth2_scheme, blacklist_token
+from ...core.db.database import async_get_db
+from ...core.exceptions.http_exceptions import UnauthorizedException
 
 router = APIRouter(tags=["login"])
