+++ src/app/core/db/crud_token_blacklist.py
@@ -1,6 +1,6 @@
-from app.crud.crud_base import CRUDBase
-from app.core.db.token_blacklist import TokenBlacklist
-from app.core.schemas import TokenBlacklistCreate, TokenBlacklistUpdate
+from ...crud.crud_base import CRUDBase
+from ..db.token_blacklist import TokenBlacklist
+from ..schemas import TokenBlacklistCreate, TokenBlacklistUpdate
 
 CRUDTokenBlacklist = CRUDBase[TokenBlacklist, TokenBlacklistCreate, TokenBlacklistUpdate, TokenBlacklistUpdate, None]
 crud_token_blacklist = CRUDTokenBlacklist(TokenBlacklist)
