+++ src/app/core/utils/rate_limit.py
@@ -3,8 +3,8 @@ from datetime import datetime
 from redis.asyncio import Redis, ConnectionPool
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.core.logger import logging
-from app.schemas.rate_limit import sanitize_path
+from ...core.logger import logging
+from ...schemas.rate_limit import sanitize_path
 
 logger = logging.getLogger(__name__)
