+++ src/app/crud/crud_base.py
@@ -13,7 +13,7 @@ from .helper import (
     _auto_detect_join_condition,
     _add_column_with_prefix
 )
-from app.core.db.database import Base
+from ..core.db.database import Base
 
 ModelType = TypeVar("ModelType", bound=Base)
 CreateSchemaType = TypeVar("CreateSchemaType", bound=BaseModel)
