+++ src/app/crud/crud_posts.py
@@ -1,6 +1,6 @@
-from app.crud.crud_base import CRUDBase
-from app.models.post import Post
-from app.schemas.post import PostCreateInternal, PostUpdate, PostUpdateInternal, PostDelete
+from .crud_base import CRUDBase
+from ..models.post import Post
+from ..schemas.post import PostCreateInternal, PostUpdate, PostUpdateInternal, PostDelete
 
 CRUDPost = CRUDBase[Post, PostCreateInternal, PostUpdate, PostUpdateInternal, PostDelete]
 crud_posts = CRUDPost(Post)
