+++ src/app/crud/crud_rate_limit.py
@@ -1,7 +1,7 @@
-from app.crud.crud_base import CRUDBase
-from app.models.rate_limit import RateLimit
+from .crud_base import CRUDBase
+from ..models.rate_limit import RateLimit
 
-from app.schemas.rate_limit import (
+from ..schemas.rate_limit import (
     RateLimitCreateInternal,
     RateLimitUpdate,
     RateLimitUpdateInternal,
