+++ src/app/crud/crud_users.py
@@ -1,6 +1,6 @@
-from app.crud.crud_base import CRUDBase
-from app.models.user import User
-from app.schemas.user import UserCreateInternal, UserUpdate, UserUpdateInternal, UserDelete
+from .crud_base import CRUDBase
+from ..models.user import User
+from ..schemas.user import UserCreateInternal, UserUpdate, UserUpdateInternal, UserDelete
 
 CRUDUser = CRUDBase[User, UserCreateInternal, UserUpdate, UserUpdateInternal, UserDelete]
 crud_users = CRUDUser(User)
