+++ src/app/models/post.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from sqlalchemy import String, DateTime, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
 
-from app.core.db.database import Base
+from ..core.db.database import Base
 
 class Post(Base):
     __tablename__ = "post"
