+++ src/app/schemas/user.py
@@ -3,7 +3,7 @@ from datetime import datetime
 
 from pydantic import BaseModel, EmailStr, Field, ConfigDict
 
-from app.core.schemas import UUIDSchema, TimestampSchema, PersistentDeletion
+from ..core.schemas import UUIDSchema, TimestampSchema, PersistentDeletion
 
 class UserBase(BaseModel):
     name: Annotated[
