+++ src/app/worker.py
@@ -4,7 +4,7 @@ from arq.connections import RedisSettings
 from arq.worker import Worker
 from typing import Any
 
-from app.core.config import settings
+from .core.config import settings
 
 asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
