+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 from fastapi.testclient import TestClient
 
-from app.main import app
+from src.app.main import app
 
 @pytest.fixture(scope="session")
 def client():
