+++ tests/helper.py
@@ -1,5 +1,4 @@
 from fastapi.testclient import TestClient
-from fastapi.routing import APIRoute
 
 def _get_token(username: str, password: str, client: TestClient):
     return client.post(
